add tmuxinator for tmux session management

+++ tmux/install.sh
@@ -1,3 +1,22 @@
 #!/bin/sh
 
-ln -s  `pwd`/tmux.conf ~/.tmux.conf
+sudo apt-get install tmuxinator
+
+export EDITOR='vim'
+
+if [ ! -f ~/.tmux.conf ]; then
+    ln -s  `pwd`/tmux.conf ~/.tmux.conf
+fi
+
+if [ ! -d ~/.tmuxinator/ ]; then
+    mkdir ~/.tmuxinator
+fi
+
+if [ ! -f ~/.tmuxinator/omegaup.yml ]; then
+    ln -s  `pwd`/omegaup.tmuxinator.yml ~/.tmuxinator/omegaup.yml
+fi
+
+if [ ! -f ~/.tmuxinator/enterpos.yml ]; then
+    ln -s  `pwd`/omegaup.enterpos.yml ~/.tmuxinator/enterpos.yml
+fi
+
